Minor fixes to software renderer, broken sky

+++ include/d_iface.h
@@ -209,8 +209,8 @@ void R_GenTile(msurface_t *psurf, void *pdest);
 #define SKYSIZE (1 << SKYSHIFT)
 #define SKYMASK (SKYSIZE - 1)
 
-extern float skyspeed, skyspeed2;
-extern float skytime;
+extern int32_t skyspeed, skyspeed2;
+extern int32_t skytime;
 
 extern int c_surf;
 extern vrect_t scr_vrect;
